+++ core/shared/src/main/scala/cron4s/datetime/DateTimeNode.scala
@@ -30,6 +30,7 @@ trait DateTimeNode[E[_ <: CronField], F <: CronField] {
     * @return true if there is a field in this date-time that matches this expression
     */
   def matchesIn[DateTime](expr: E[F], DT: IsDateTime[DateTime]): Predicate[DateTime] = Predicate { dt =>
+    import cats.syntax.either._
     val current = DT.get(dt, expr.unit.field)
     current.map(expr.matches).getOrElse(false)
   }
@@ -68,7 +69,7 @@ trait DateTimeNode[E[_ <: CronField], F <: CronField] {
   def stepIn[DateTime](expr: E[F], DT: IsDateTime[DateTime])(dateTime: DateTime, step: Int): Option[DateTime] = {
     import cats.syntax.either._
     for {
-      current  <- DT.get(dateTime, expr.unit.field)
+      current  <- DT.get(dateTime, expr.unit.field).toOption
       newValue <- expr.step(current, step).map(_._1)
       adjusted <- DT.set(dateTime, expr.unit.field, newValue).toOption
     } yield adjusted