+++ core/shared/src/main/scala/cron4s/datetime/errors.scala
@@ -16,8 +16,15 @@
 
 package cron4s.datetime
 
+import cats.Eq
 import cron4s.CronField
 
-sealed trait DateTimeError
-final case class UnsupportedField(field: CronField) extends DateTimeError
-final case class InvalidFieldValue(field: CronField, value: Int) extends DateTimeError
+sealed abstract class DateTimeError(msg: String) extends Exception(msg)
+object DateTimeError {
+  implicit val dateTimeErrorEq: Eq[DateTimeError] = Eq.fromUniversalEquals
+}
+
+final case class UnsupportedField(field: CronField)
+  extends DateTimeError(s"Field $field is not supported")
+final case class InvalidFieldValue(field: CronField, value: Int)
+  extends DateTimeError(s"Value $value is not valid for field $field")