+++ project/Dependencies.scala
@@ -2,12 +2,13 @@ import sbt._
 import Keys._
 
 import sbtcrossproject.CrossPlugin.autoImport._
-import org.scalajs.sbtplugin.ScalaJSPlugin.autoImport.scalaJSVersion
+import scalajscrossproject.ScalaJSCrossPlugin.autoImport.{toScalaJSGroupID => _, _}
+import org.scalajs.sbtplugin.ScalaJSPlugin.autoImport._
 
 object Dependencies {
 
   object version {
-    val cats        = "1.0.0-RC1"
+    val cats        = "1.0.0-RC2"
     val shapeless   = "2.3.2"
     val fastparse   = "1.0.0"
     val scalacheck  = "1.13.5"
@@ -16,13 +17,13 @@ object Dependencies {
     val catalysts   = "0.0.5"
 
     val jodaTime      = "2.9.9"
-    val jodaConvert   = "1.8.3"
+    val jodaConvert   = "1.9.2"
     val momentjs      = "0.8.1"
     val scalaJavaTime = "2.0.0-M12"
   }
 
   val macroParadise = compilerPlugin("org.scalamacros" % "paradise"       % "2.1.1" cross CrossVersion.full)
-  val kindProjector = compilerPlugin("org.spire-math"  % "kind-projector" % "0.9.4" cross CrossVersion.binary)
+  val kindProjector = compilerPlugin("org.spire-math"  % "kind-projector" % "0.9.5" cross CrossVersion.binary)
   lazy val compilerPlugins = Seq(macroParadise, kindProjector)
 
   lazy val core = Def.settings {